refactor(geocoding): refactoring geocode.services && add nowatchtest npm script
closes #12
