+++ src/app/services/geocode.services/geocode.ts
@@ -4,4 +4,3 @@ export interface Geocode {
   getReverseGeocoding(lat: number, long: number): any;
 }
 
-export const MAPZEN_API_KEY = 'mapzen-YJqMyuT';
